+++ mcu/nrf52832/gpio.c
@@ -9,27 +9,27 @@
 
 #include <mcu/common.h>
 
-bool gpio_get(struct gpio *bank, int pin)
+bool gpio_get(struct gpio *bank, uint8_t pin)
 {
 	return !!(bank->IN & (1 << pin));
 }
 
-void gpio_set(struct gpio *bank, int pin)
+void gpio_set(struct gpio *bank, uint8_t pin)
 {
 	bank->OUT |= (1 << pin);
 }
 
-void gpio_clr(struct gpio *bank, int pin)
+void gpio_clr(struct gpio *bank, uint8_t pin)
 {
 	bank->OUT &= ~(1 << pin);
 }
 
-void gpio_toggle(struct gpio *bank, int pin)
+void gpio_toggle(struct gpio *bank, uint8_t pin)
 {
 	bank->OUT ^= (1 << pin);
 }
 
-void gpio_in(struct gpio *bank, int pin)
+void gpio_in(struct gpio *bank, uint8_t pin)
 {
 	/* Connect input buffer */
 	bank->PIN_CNF[pin] &= ~(1 << 1);
@@ -37,7 +37,7 @@ void gpio_in(struct gpio *bank, int pin)
 	bank->DIR &= ~(1 << pin);
 }
 
-void gpio_out(struct gpio *bank, int pin)
+void gpio_out(struct gpio *bank, uint8_t pin)
 {
 	/* Disconnect input buffer */
 	bank->PIN_CNF[pin] |= (1 << 1);
